This is a new tool to generate EncodeRLP (and DecodeRLP) method implementations from a struct definition.

I have applied this tool to several types in core/types, and here are the benchmark results:

```
name                             old time/op    new time/op    delta
EncodeRLP/legacy-header-8           328ns ± 0%     237ns ± 1%   -27.63%  (p=0.000 n=8+8)
EncodeRLP/london-header-8           353ns ± 0%     247ns ± 1%   -30.06%  (p=0.000 n=8+8)
EncodeRLP/receipt-for-storage-8     237ns ± 0%     123ns ± 0%   -47.86%  (p=0.000 n=8+7)
EncodeRLP/receipt-full-8            297ns ± 0%     301ns ± 1%    +1.39%  (p=0.000 n=8+8)

name                             old speed      new speed      delta
EncodeRLP/legacy-header-8        1.66GB/s ± 0%  2.29GB/s ± 1%   +38.19%  (p=0.000 n=8+8)
EncodeRLP/london-header-8        1.55GB/s ± 0%  2.22GB/s ± 1%   +42.99%  (p=0.000 n=8+8)
EncodeRLP/receipt-for-storage-8  38.0MB/s ± 0%  64.8MB/s ± 0%   +70.48%  (p=0.000 n=8+7)
EncodeRLP/receipt-full-8          910MB/s ± 0%   897MB/s ± 1%    -1.37%  (p=0.000 n=8+8)

name                             old alloc/op   new alloc/op   delta
EncodeRLP/legacy-header-8           0.00B          0.00B           ~     (all equal)
EncodeRLP/london-header-8           0.00B          0.00B           ~     (all equal)
EncodeRLP/receipt-for-storage-8     64.0B ± 0%      0.0B       -100.00%  (p=0.000 n=8+8)
EncodeRLP/receipt-full-8             320B ± 0%      320B ± 0%      ~     (all equal)
```