@holiman 2022-01-25
triage discussion: this could be fuzzed. We could create types, like `headerAlt`, which is an exact replica of `types.Header`. Then instantiate a `hdrA headerAlt` and `hdrB types.Header`, populate them identically, and encode - one via reflection and one via generated code, and then we can compare the outputs.