@holiman 2022-01-25
Triage discussion: I'll run this a bit, but also apply the panic on encoding value-types